feat: add onClick to MenuItem
Adds an optional `onClick` lambda to the `MenuItem` class. This allows for defining click behavior directly within the menu item
 definition.

The `EditorTopBarMenuAdapter` is updated to utilize the `onClick` lambda when a menu item is clicked.

The `EditorActivity` now defines click behavior for each menu item using the `onClick` lambda, replacing the previous `when` statement.